CRUD API WITH Authentication

+++ src/main/java/com/kimmyungho/board/service/PostService.java
@@ -1,6 +1,8 @@
 package com.kimmyungho.board.service;
 
 import com.kimmyungho.board.exception.post.PostNotFoundException;
+import com.kimmyungho.board.exception.user.UserNotAllowedException;
+import com.kimmyungho.board.model.entity.UserEntity;
 import com.kimmyungho.board.model.post.Post;
 import com.kimmyungho.board.model.post.PostPathRequestBody;
 import com.kimmyungho.board.model.post.PostPostRequestBody;
@@ -38,14 +40,15 @@ PostService {
         return Post.from(postEntity);
     }
 
-    public Post createPost(PostPostRequestBody postPostRequestBody) {
-        var postEntity = new PostEntity();
-            postEntity.setBody(postPostRequestBody.getBody());
-        var savedPostEntity = postEntityRepository.save(postEntity);
+    public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentuser) {
+
+        var savedPostEntity = postEntityRepository.save(
+                PostEntity.of(postPostRequestBody.getBody(), currentuser)
+        );
         return Post.from(savedPostEntity);
     }
 
-    public Post updatePost(Long postId, PostPathRequestBody postPathRequestBody) {
+    public Post updatePost(Long postId, PostPathRequestBody postPathRequestBody, UserEntity currentuser) {
         var postEntity =
                 postEntityRepository
                         .findById(postId)
@@ -53,19 +56,25 @@ PostService {
                                 () -> new PostNotFoundException(postId)
                         );
 
+        if (! postEntity.getUser().equals(currentuser)) {
+            throw new UserNotAllowedException();
+        }
         postEntity.setBody(postPathRequestBody.getBody());
         var updatePostEntity = postEntityRepository.save(postEntity);
 
         return Post.from(updatePostEntity);
     }
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, UserEntity currentuser) {
         var postEntity =
                 postEntityRepository
                         .findById(postId)
                         .orElseThrow(
                                 () -> new PostNotFoundException(postId)
                         );
+        if (! postEntity.getUser().equals(currentuser)) {
+            throw new UserNotAllowedException();
+        }
         postEntityRepository.delete(postEntity);
     }
 }
